feat: config.tokenizer
Introduces a way to extend/customize the tokenizer and its provided methods (setup, tokenize).

- Abstract prism tokenizer
- Adds package export `syntax-highlight-element/tokenizer/prism`
- Export  `setupTokenHighlights` method
- Flat merge config with defaultConfig
