+++ index.html
@@ -32,12 +32,17 @@
     <script>
       window.she = window.she || {};
       window.she.config = {
-        languages: ['html', 'css', 'js', 'jsx', 'md'], // Default
+        languages: ['html', 'css', 'js', 'jsx', 'md'],
         // Optional: language specific token type overwrites
         languageTokens: {
           css: ['important'],
           md: ['title', 'list'],
-        }
+        },
+        // Optional: extend/customize tokenizer
+        // tokenizer: {
+        //   async setup() {},
+        //   tokenize: () => [],
+        // }
       }
     </script>
   </head>
@@ -87,15 +92,55 @@ import SyntaxHighlightElement from 'syntax-highlight-element';</syntax-highlight
 </syntax-highlight>
 
     <script type="module">
-      // Manually loading prism dependencies
-      // import 'https://cdn.jsdelivr.net/npm/prismjs@1.30.0/components/prism-core.min.js'
+      // /**
+      //  * Option 1. autoload - default
+      //  */
+      // import '/src/index.js'
+      // import SyntaxHighlightElement from '/src/index.js'
+
+      /**
+       * Option 2. Opt out of prism autoload by manually importing prism core
+       */
+      import 'https://cdn.jsdelivr.net/npm/prismjs@1.30.0/components/prism-core.min.js'
+      import SyntaxHighlightElement from '/src/index.js?define=false'
+
+      /**
+       * Option 2.a. Load lang deps manually (without dependency resolution)
+       */
+      // // - CDN
       // import 'https://cdn.jsdelivr.net/npm/prismjs@1.30.0/components/prism-markup.min.js'
       // import 'https://cdn.jsdelivr.net/npm/prismjs@1.30.0/components/prism-css.min.js'
       // import 'https://cdn.jsdelivr.net/npm/prismjs@1.30.0/components/prism-clike.min.js'
       // import 'https://cdn.jsdelivr.net/npm/prismjs@1.30.0/components/prism-javascript.min.js'
+      // import 'https://cdn.jsdelivr.net/npm/prismjs@1.30.0/components/prism-jsx.min.js'
       // import 'https://cdn.jsdelivr.net/npm/prismjs@1.30.0/components/prism-markdown.min.js'
+      // // - npm
+      // import 'prismjs/components/prism-markup'
+      // import 'prismjs/components/prism-css'
+      // import 'prismjs/components/prism-clike'
+      // import 'prismjs/components/prism-javascript'
+      // import 'prismjs/components/prism-jsx'
+      // import 'prismjs/components/prism-markdown'
 
-      import '/src/index.js'
+      /**
+       * Option 2.b. resolve language dependencies and load from URL/CDN
+       */
+      // - CDN
+      import { loadPrismLanguage } from '/src/tokenizer/prism.js'
+      await loadPrismLanguage({
+        baseUrl: 'https://unpkg.com/prismjs@1.30.0',
+        language: window.she.config.languages
+      });
+      // // - npm
+      // import { resolveLanguageDependencies } from '/src/tokenizer/prism.js'
+      // const langDeps = resolveLanguageDependencies(window.she.config.languages);
+      // for await (const lang of langDeps) {
+      //   const importPath = `prismjs/components/prism-${lang}`
+      //   await import(importPath);
+      //   console.log({importPath})
+      // }
+      // // - Manually define SyntaxHighlightElement with ?define=false once all dependecies are loaded
+      window.SyntaxHighlightElement = SyntaxHighlightElement.define();
     </script>
     <script>
       document.addEventListener('DOMContentLoaded', () => {
