+++ src/constants.js
@@ -1,11 +1,21 @@
+import { tokenizer as defaultTokenizer } from './tokenizer/prism';
+
 export const NAMESPACE = 'she';
 
 /**
  * @typedef Config
  * @type {object}
- * @property {string[]} languages - Prism languages.
+ * @property {string[]} languages - Languages.
  * @property {{ [key: string]: string[] }} languageTokens - Language specific token types.
+ * @property {object} tokenizer - Tokenizer.
  */
 
 /** @type {Config} */
-export const CONFIG = window[NAMESPACE]?.config || {};
+export const CONFIG = Object.assign(
+  {
+    languages: ['markup', 'css', 'javascript'],
+    languageTokens: {},
+    tokenizer: Object.assign(defaultTokenizer, window[NAMESPACE]?.config?.tokenizer || {}),
+  },
+  window[NAMESPACE]?.config || {},
+);
